refactor: migrate no-import-module-exports rule
